Migration process: uint256 main methods removed, using arith_uint256 for now.
Once all of the invalidly instantiated fields using uint256 as a number get migrated to the arith_uint256, the uint256 file will be replaced with the blob_uint256 file.
